Refactored -> RobotCore
Made everything go through RobotCore. 
All the methods are best fit as it is, but we may want to discuss some of them as I have made getWinCell and getLoseCell automatically make its move through  the methods, and some other questionable stuff. But this is organized and something to build upon. Kept everything else as is.