Seperate Game Logic from UI
A beginning of issues #98 and #99 and some other issues we haven't currently  made. Made a Grid that reads a tmx map file and puts each layer in the map into a gridlayer with GridPoints as key and TileName as value for each position in the various layers containing an element (for example a flag in the flag layer).