#144 #149 #150  Cleaning register when manually running the game. Fixed problem 'test 1' and 'test 3' from #150.
The register is always cleaned at the end of a 'round'. It is crucially important that the 'cleanRegister' is run in order for the register to be correctly setup.
When we do manuel tests, like moving robots, pressing 'T' to shoot etc. We have to press 'E' after we are done to run the final stages of a round: namely cleanup and restore robots. Otherwise the game logic will be broken.
If we run games normally, by only playing cards. This is not an issue.